Moved scans views into blueprint, moved github utils into time_bp
