+++ dashboard/forms.py
@@ -40,35 +40,6 @@ class StudyOverviewForm(FlaskForm):
     study_id = HiddenField()
 
 
-class SliceTimingForm(FlaskForm):
-    timings = TextAreaField('NewTimings',
-                            id="new_timings",
-                            render_kw={
-                                'rows':
-                                4,
-                                'cols':
-                                65,
-                                'required':
-                                True,
-                                'placeholder':
-                                "Enter comma " + "separated slice " + "timings"
-                            })
-    submit = SubmitField('Update', id='submit_timings')
-
-
-class ScanChecklistForm(FlaskForm):
-    comment = TextAreaField('Comment:',
-                            id='scan-comment',
-                            validators=[DataRequired()],
-                            render_kw={
-                                'placeholder': 'Add description',
-                                'rows': 12,
-                                'required': True,
-                                'maxlength': '1028'
-                            })
-    submit = SubmitField('Submit')
-
-
 class AnalysisForm(FlaskForm):
     name = TextField('Brief name', validators=[DataRequired()])
     description = TextAreaField('Description', validators=[DataRequired()])
