+++ dashboard/main/views.py
@@ -17,12 +17,9 @@ from ..queries import (query_metric_values_byid, query_metric_types,
                       query_metric_values_byname, find_subjects, find_sessions,
                       find_scans)
 from ..models import Study, Site, User, Timepoint, Analysis, AccountRequest
-from ..forms import (SelectMetricsForm, StudyOverviewForm, ScanChecklistForm,
-                     AnalysisForm, SliceTimingForm)
+from ..forms import (SelectMetricsForm, StudyOverviewForm, AnalysisForm)
 from ..view_utils import (report_form_errors, get_timepoint, get_scan,
-                         get_redcap_record,
-                         dashboard_admin_required, study_admin_required,
-                         prev_url, is_safe_url)
+                          prev_url, is_safe_url)
 from ..exceptions import InvalidUsage
 
 logger = logging.getLogger(__name__)
@@ -106,10 +103,10 @@ def search_data(search_string=None):
         study = scans[0].session.timepoint.accessible_study(current_user)
         if study:
             return redirect(
-                url_for('main.scan', study_id=study.id, scan_id=scans[0].id))
+                url_for('scans.scan', study_id=study.id, scan_id=scans[0].id))
 
     scans = [
-        url_for('main.scan',
+        url_for('scans.scan',
                 study_id=scan.session.timepoint.accessible_study(current_user),
                 scan_id=scan.id) for scan in scans
         if scan.session.timepoint.accessible_study(current_user)
@@ -120,131 +117,6 @@ def search_data(search_string=None):
                            subjects=subjects,
                            sessions=sessions,
                            scans=scans)
-
-
-@main.route('/study/<string:study_id>/scan/<int:scan_id>',
-           methods=['GET', 'POST'])
-@login_required
-def scan(study_id, scan_id):
-    scan = get_scan(scan_id,
-                    study_id,
-                    current_user,
-                    fail_url=url_for('main.study', study_id=study_id))
-    checklist_form = ScanChecklistForm(obj=scan.get_checklist_entry())
-    slice_timing_form = SliceTimingForm()
-    return render_template('scan/main.html',
-                           scan=scan,
-                           study_id=study_id,
-                           checklist_form=checklist_form,
-                           slice_timing_form=slice_timing_form)
-
-
-@main.route('/study/<string:study_id>/papaya/<int:scan_id>', methods=['GET'])
-@login_required
-def papaya(study_id, scan_id):
-    scan = get_scan(scan_id,
-                    study_id,
-                    current_user,
-                    fail_url=url_for('main.study', study_id=study_id))
-    name = os.path.basename(utils.get_nifti_path(scan))
-    return render_template('scan/viewer.html',
-                           study_id=study_id,
-                           scan_id=scan_id,
-                           nifti_name=name)
-
-
-@main.route('/study/<string:study_id>/slice-timing/<int:scan_id>',
-           methods=['POST'])
-@main.route('/study/<string:study_id>/slice-timing/<int:scan_id>/auto/<auto>',
-           methods=['GET'])
-@main.route('/study/<string:study_id>/slice-timing/<int:scan_id>'
-           '/delete/<delete>')
-@login_required
-def fix_slice_timing(study_id, scan_id, auto=False, delete=False):
-    dest_url = url_for('main.scan', study_id=study_id, scan_id=scan_id)
-
-    scan = get_scan(scan_id, study_id, current_user)
-    # Need a new dictionary to get the changes to actually save
-    new_json = dict(scan.json_contents)
-
-    if auto:
-        new_json["SliceTiming"] = scan.get_header_diffs(
-        )["SliceTiming"]["expected"]
-    elif delete:
-        del new_json["SliceTiming"]
-    else:
-        timing_form = SliceTimingForm()
-        if not timing_form.validate_on_submit():
-            flash("Failed to update slice timings")
-            return redirect(dest_url)
-
-        new_timings = timing_form.timings.data
-        new_timings = new_timings.replace("[", "").replace("]", "")
-        new_json["SliceTiming"] = [
-            float(item.strip()) for item in new_timings.split(",")
-        ]
-
-    try:
-        utils.update_json(scan, new_json)
-    except Exception as e:
-        logger.error("Failed updating slice timings for scan {}. Reason {} "
-                     "{}".format(scan_id,
-                                 type(e).__name__, e))
-        flash("Failed during slice timing update. Please contact an admin for "
-              "help")
-        return redirect(dest_url)
-
-    utils.update_header_diffs(scan)
-    flash("Update successful")
-
-    return redirect(dest_url)
-
-
-@main.route('/study/<string:study_id>/scan/<int:scan_id>/review',
-           methods=['GET', 'POST'])
-@main.route('/study/<string:study_id>/scan/<int:scan_id>/review/<sign_off>',
-           methods=['GET', 'POST'])
-@main.route('/study/<string:study_id>/scan/<int:scan_id>/delete/<delete>',
-           methods=['GET', 'POST'])
-@main.route('/study/<string:study_id>/scan/<int:scan_id>/update/<update>',
-           methods=['GET', 'POST'])
-@login_required
-def scan_review(study_id, scan_id, sign_off=False, delete=False, update=False):
-    scan = get_scan(scan_id,
-                    study_id,
-                    current_user,
-                    fail_url=url_for('main.study', study_id=study_id))
-    dest_url = url_for('main.scan', study_id=study_id, scan_id=scan_id)
-
-    if delete:
-        entry = scan.get_checklist_entry()
-        entry.delete()
-        return redirect(dest_url)
-
-    if sign_off:
-        # Just in case the value provided in the URL was not boolean
-        sign_off = True
-
-    checklist_form = ScanChecklistForm()
-    if checklist_form.is_submitted():
-        if not checklist_form.validate_on_submit():
-            report_form_errors(checklist_form)
-            return redirect(dest_url)
-        comment = checklist_form.comment.data
-    else:
-        comment = None
-
-    if update:
-        # Update is done separately so that a review entry can't accidentally
-        # be changed from 'flagged' to blacklisted.
-        if comment is None:
-            flash("Cannot update entry with empty comment")
-            return redirect(dest_url)
-        scan.add_checklist_entry(current_user.id, comment)
-        return redirect(dest_url)
-
-    scan.add_checklist_entry(current_user.id, comment, sign_off)
-    return redirect(url_for('main.scan', study_id=study_id, scan_id=scan_id))
 
 
 @main.route('/study/<string:study_id>', methods=['GET', 'POST'])
@@ -599,24 +471,3 @@ def static_qc_page(study_id,
         qc_dir, _ = os.path.split(timepoint.static_page)
         return send_from_directory(qc_dir, image)
     return send_file(timepoint.static_page)
-
-
-# The file name (with correct extension) needs to be last part of the URL or
-# papaya will fail to read the file because it wont figure out on its own
-# whether or not a file needs decompression
-@main.route('/study/<string:study_id>/load_scan/<int:scan_id>/'
-           '<string:file_name>')
-@login_required
-def load_scan(study_id, scan_id, file_name):
-    scan = get_scan(scan_id, study_id, current_user, fail_url=prev_url())
-    full_path = utils.get_nifti_path(scan)
-    try:
-        result = send_file(full_path,
-                           as_attachment=True,
-                           attachment_filename=file_name,
-                           mimetype="application/gzip")
-    except IOError as e:
-        logger.error("Couldnt find file {} to load scan view for user "
-                     "{}".format(full_path, current_user))
-        result = not_found_error(e)
-    return result
