+++ dashboard/templates/scan/viewer.html
@@ -4,7 +4,7 @@
     <script type="text/javascript">
         var params = [];
         params["worldSpace"] = true;
-        params["images"] = ["{{ url_for('main.load_scan', study_id=study_id, scan_id=scan_id, file_name=nifti_name) }}"];
+        params["images"] = ["{{ url_for('scans.load_scan', study_id=study_id, scan_id=scan_id, file_name=nifti_name) }}"];
         params["mainView"] = 'coronal';
         params["interpolation"] = true;
     </script>
