+++ dashboard/timepoints/templates/sessions/scan_table.html
@@ -16,11 +16,11 @@
           <td class="scan_series"> {{ scan.series }}</td>
           <td class="scan_type"> {{ scan.tag }}</td>
           <td class="scan_desc">
-            <a href="{{ url_for('main.scan', study_id=study_id, scan_id=scan.id ) }}">
+            <a href="{{ url_for('scans.scan', study_id=study_id, scan_id=scan.id ) }}">
               {{ scan.description }}
             </a>
             {% if scan.source_id %}
-              <a href="{{ url_for('main.scan', study_id=scan.source_data.get_study().id, scan_id=scan.source_id) }}">
+              <a href="{{ url_for('scans.scan', study_id=scan.source_data.get_study().id, scan_id=scan.source_id) }}">
                 <span class="badge" title="Click to visit source scan">
                   linked scan
                 </span>
