+++ dashboard/timepoints/utils.py
@@ -22,6 +22,64 @@ def search_issues(token, timepoint):
     return result
 
 
+def handle_issue(token, issue_form, study_id, timepoint):
+    title = clean_issue_title(issue_form.title.data, timepoint)
+    study = Study.query.get(study_id)
+
+    staff_member = study.choose_staff_contact()
+    if staff_member:
+        assigned_user = staff_member.username
+    else:
+        assigned_user = None
+
+    try:
+        issue = make_issue(token, title, issue_form.body.data,
+                           assign=assigned_user)
+    except Exception as e:
+        logger.error("Failed to create a GitHub issue for {}. "
+                     "Reason: {}".format(timepoint, e))
+        flash("Failed to create issue '{}'".format(title))
+    else:
+        flash("Issue '{}' created!".format(title))
+
+
+def clean_issue_title(title, timepoint):
+    title = title.rstrip()
+    if not title:
+        title = timepoint
+    elif title.endswith('-'):
+        title = title[:-1].rstrip()
+    elif timepoint not in title:
+        title = timepoint + " - " + title
+    return title
+
+
+def make_issue(token, title, body, assign=None):
+    try:
+        repo = get_issues_repo(token)
+        if assign:
+            issue = repo.create_issue(title, body, assignee=assign)
+        else:
+            # I thought a default of None would be a clever way to avoid
+            # needing an if/else here but it turns out 'assignee' will raise a
+            # mysterious exception when set to None :( So... here we are
+            issue = repo.create_issue(title, body)
+    except Exception as e:
+        raise Exception("Can't create new issue '{}'. Reason: {}".format(
+            title, e))
+    return issue
+
+
+def get_issues_repo(token):
+    owner = current_app.config['GITHUB_OWNER']
+    repo = current_app.config['GITHUB_REPO']
+    try:
+        repo = Github(token).get_user(owner).get_repo(repo)
+    except Exception as e:
+        raise Exception("Can't retrieve github issues repo. {}".format(e))
+    return repo
+
+
 def delete_timepoint(timepoint):
     config = datman.config.config(study=timepoint.get_study().id)
